Fixing unified memory and environment measurement The CUPTI unified memory and environment measurements had bogus timestamps (not scaled right) and weren't going to the right threads.
Former-commit-id: 10be98c461b8b3bea845f1515fac540442277151

+++ include/Profile/CuptiActivity.h
@@ -135,7 +135,7 @@ extern "C" void * Tau_return_context_userevent(const char *name);
 extern "C" void metric_set_gpu_timestamp(int tid, double value);
 
 extern "C" void Tau_cupti_find_context_event(
-						TauContextUserEvent** u, 
+						TauContextUserEvent** u,
 						const char *name,
             bool context);
 
@@ -182,7 +182,7 @@ extern "C" void Tau_cupti_register_memcpy_event(
 						double stop,
 						int bytes_copied,
 						int memcpy_type,
-						int direction, 
+						int direction,
 						int taskId);
 
 extern "C" void Tau_cupti_register_unifmem_event(
@@ -310,13 +310,13 @@ static std::set<uint32_t> set_gpuThread;
 
 int gpu_occupancy_available(int deviceId);
 
-void record_gpu_occupancy(int32_t blockX, 
+void record_gpu_occupancy(int32_t blockX,
                           int32_t blockY,
                           int32_t blockZ,
 			  uint16_t registersPerThread,
 			  int32_t staticSharedMemory,
                           uint32_t deviceId,
-                          const char *name, 
+                          const char *name,
                           eventMap_t *map);
 
 void record_gpu_launch(int cId, const char *name);
@@ -325,7 +325,7 @@ void record_imix_counters(const char* name, uint32_t taskId, uint32_t streamId, 
 void transport_imix_counters(uint32_t vec, Instrmix imixT, const char* name, uint32_t taskId, uint32_t streamId, uint32_t contextId, uint32_t id, uint64_t end, TauContextUserEvent * tc);
 ImixStats write_runtime_imix(uint32_t corrId, uint32_t taskId, std::map<std::pair<int, int>, CudaOps> map_disassem, std::string kernel);
 void write_sass_counters();
-void record_environment_counters(const char *name, uint32_t taskId, uint32_t device, uint32_t streamId, uint32_t contextId, uint32_t id, uint32_t end);
+void record_environment_counters(const char *name, uint32_t taskId, uint32_t device, uint32_t streamId, uint32_t contextId, uint32_t id, uint64_t end);
 void transport_environment_counters(std::vector<uint32_t> vec, EnvType envT, const char* name, uint32_t taskId, uint32_t deviceId, uint32_t streamId, uint32_t contextId, uint32_t id, uint64_t end, TauContextUserEvent* tc);
 
 int get_device_count();
@@ -376,7 +376,7 @@ typedef std::map<uint32_t, CUpti_ActivityDevice> device_map_t;
 //static std::map<uint32_t, CUpti_ActivityDevice> deviceMap;
 
 void record_gpu_counters_at_launch(int device, int task);
-void record_gpu_counters_at_sync();  
+void record_gpu_counters_at_sync();
 void record_gpu_counters_at_sync(int device, int task);
 
 void clear_counters(int device);
@@ -403,6 +403,6 @@ int get_task_from_id(int id, int task);
 	id++
 #endif
 
-	//Tau_metadata("GPU " SX(name), str_##name.str().c_str()); 
+	//Tau_metadata("GPU " SX(name), str_##name.str().c_str());
 
 #endif //CUPTI_ACTIVITY_H
